requirements.txt
pandas
python-dotenv
langchain
langchain_community
langchain_openai
pysqlite3-binary
scikit-learn